Feat : Placeholder Utils로 분리
