+++ Cookiee/Cookiee/Presentation/Event/EventAddView.swift
@@ -141,18 +141,6 @@ struct InitialAddMessageCardView: View {
     }
 }
 
-extension View {
-    func placeholder<Content: View>(
-        when shouldShow: Bool,
-        @ViewBuilder placeholder: () -> Content) -> some View {
-
-        ZStack(alignment: .leading) {
-            placeholder().opacity(shouldShow ? 1 : 0)
-            self
-        }
-    }
-}
-
 #Preview {
     EventAddView()
 }
