+++ Cookiee/Cookiee/Utils/Placeholder.swift
@@ -1,0 +1,20 @@
+//
+//  Placeholder.swift
+//  Cookiee
+//
+//  Created by minseo Kyung on 10/1/24.
+//
+
+import SwiftUI
+
+extension View {
+    func placeholder<Content: View>(
+        when shouldShow: Bool,
+        @ViewBuilder placeholder: () -> Content) -> some View {
+
+        ZStack(alignment: .leading) {
+            placeholder().opacity(shouldShow ? 1 : 0)
+            self
+        }
+    }
+}
